backend/requirements.txt
# Core Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.1

# Validation & Settings
pydantic==2.10.4
pydantic-settings==2.7.1
pydantic[email]

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Redis
redis[hiredis]==5.2.1

# AWS
boto3==1.36.4

# HTTP Client (for webhooks, shipping APIs)
httpx==0.28.1

# Image Processing (for Lambda, optional local)
Pillow==11.1.0

# PDF Generation
weasyprint==63.1

# Background Jobs
celery[redis]==5.4.0

# Utilities
python-slugify==8.0.4
phonenumbers==8.13.52

# Testing
pytest==8.3.4
pytest-asyncio==0.25.2
httpx==0.28.1

# Linting
ruff==0.9.3
